+++ packages/ethereum-ethers-v6/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @irys/web-upload-ethereum-ethers-v6
 
+## 0.0.2
+
+### Patch Changes
+
+- Fix use of export paths for TypeScript
+
+- Updated dependencies []:
+  - @irys/web-upload-ethereum@0.0.2
+  - @irys/web-upload@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes
@@ -10,4 +20,3 @@
   - @irys/web-upload-ethereum@0.0.1
   - @irys/upload-core@0.0.1
   - @irys/web-upload@0.0.1
-