+++ packages/ethereum-web/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @irys/web-upload-ethereum
 
+## 0.0.2
+
+### Patch Changes
+
+- Fix use of export paths for TypeScript
+
+- Updated dependencies []:
+  - @irys/web-upload@0.0.2
+
 ## 0.0.1
 
 ### Patch Changes
@@ -9,4 +18,3 @@
 - Updated dependencies []:
   - @irys/upload-core@0.0.1
   - @irys/web-upload@0.0.1
-