+++ packages/upload-node/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @irys/upload
 
+## 0.0.2
+
+### Patch Changes
+
+- Fix use of export paths for TypeScript
+
 ## 0.0.1
 
 ### Patch Changes
@@ -8,4 +14,3 @@
 
 - Updated dependencies []:
   - @irys/upload-core@0.0.1
-